ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
